[Kernel] bug fix in Fat file system, XHCI hot plug support and Chitralekha Widgets
xhci lack support of hot plugging, now a USB thread is spawned in order to
handle USB hot plugging messages, and initialise the device and attacha a
driver to it.
Chitralekha onoff and slider widget was lacking ActionHandler callback, now
whenever the state of the widget is changed it'll call the associated action
handler from its user application

Thank you,
Manas Kamal Choudhury [08-04-2025, 4:26 P.M]

+++ BaseHdr/Drivers/usb.h
@@ -200,8 +200,8 @@ typedef int (*poll_wait_callback)(_au_usb_dev_* dev, int poll_wait);
 #pragma pack(push,1)
 typedef struct _au_usb_dev_ {
 	void* data;
-	uint8_t classCode;
-	uint8_t subClassCode;
+	uint32_t classCode;
+	uint32_t subClassCode;
 	uint16_t vendorID;
 	uint16_t deviceID;
 	int configValue;
